requirements.txt
fastapi
uvicorn[standard]
numpy
pypdf
sentence-transformers
pinecone-client
pinecone
transformers
torch
python-multipart
accelerate
python-dotenv
